Create temp_general_comments
Just general comments about the project. Not sure how/where best to put them at this stage. 

Each country/player should have the following: 
An exact population number, every member of which has an original identity, made available if queried (by players with such access). Each citizen of a country has an ID number, age, education, wealth, technical skill-level, health, geographical home, geographical current location.
 (legal relationship status? 2 generations of ancestory info?).  Of course, this would be too much data to generate for every person at onset, but this data should be created upon a query about an individual and after that stick with the individual. It could be generated randomly with probability biases. Name generation could be random. Education, age, wealth, industry, could, be probabilistic with respect to overall traits for the population.   

In theory, at its most advanced version, Economic Warfare could be a game in which individual players are not command & control dictators of the country, but could be individual citizens, with freedom of movement, association, etc. In general, however, this will not be the case, individual citizens will be automatons acting according to some basic laws of behavior: economic requirements. Only if/when tens of thousands of players are involved should players be expected to take roles as individual citizens. But it would be really cool if that were to occur! 

With this in mind, the game should start with a high-level version, but have the ability to expand to very complex low-level activity. The handling of Game Time would become a big issue if this occurs but let's burn that bridge if when we cross it. 

So we start at the macro level and advance to the micro. In the early version of the game, I picture each country with the basics of: 
A country with borders
A local currency with maybe a default reserve of foreign currencies (or no foreign currency reserve yet)
A set amount of natural resources: arable land better/worse for particular crops, a population than can work, go to war, and/or be provided for, mining and manufacturing industries (how detailed these inputs and outputs are will come down to how much the game expands over time);
 
A big part of the action of the game will come down to negotiating with other players/countries. Who to engage in trade with? Whom to ally with, whom to invade. Thus, players should be able to privately text other players in order to agree with and deceive each other.